Added hardcoded Shop with soft currency items

+++ Assets/Scripts/Services/MetaResources/IMetaResourcesService.cs
@@ -1,4 +1,4 @@
 ﻿public interface IMetaResourcesService : IService
 {
-    int PlayerMoney { get; set; }
+    float PlayerMoney { get; set; }
 }
